examples/auxo/requirements.txt
tensorboard
numba==0.48.0
pip==20.0.2
torch_optimizer
torch
tensorflow
torchvision
transformers
scipy==1.4.1
matplotlib==3.1.3
torch_baidu_ctc==0.3.0
tensorboardX==2.1
overrides==3.1.0
python-levenshtein==0.12.0
pandas==1.1.0
PyYAML
pytest
sox==1.3.7
grpcio==1.40.0
gym
jupyter
pillow==9
sentencepiece
gdown
h5py
librosa==0.7.2
SoundFile
resampy==0.3.1
kubernetes
wandb
nltk

